charts refactoring: primitive attributes replaced by Object type atts; default attribute values moved to chart themes; new pseudo serialization constant added
git-svn-id: http://code.jaspersoft.com/svn/repos/jasperreports/trunk/jasperreports@2469 76e25fc1-081a-0410-bb1b-c96bc70638b7

+++ src/net/sf/jasperreports/engine/xml/JRChartPlotFactory.java
@@ -62,15 +62,15 @@ public class JRChartPlotFactory extends JRBaseFactory
 
 		String foregroundAlpha = atts.getValue(JRXmlConstants.ATTRIBUTE_foregroundAlpha);
 		if (foregroundAlpha != null && foregroundAlpha.length() > 0)
-			plot.setForegroundAlpha(Float.valueOf(foregroundAlpha).floatValue());
+			plot.setForegroundAlpha(Float.valueOf(foregroundAlpha));
 
 		String backgroundAlpha = atts.getValue(JRXmlConstants.ATTRIBUTE_backgroundAlpha);
 		if (backgroundAlpha != null && backgroundAlpha.length() > 0)
-			plot.setBackgroundAlpha(Float.valueOf(backgroundAlpha).floatValue());
+			plot.setBackgroundAlpha(Float.valueOf(backgroundAlpha));
 
 		String labelRotation = atts.getValue(JRXmlConstants.ATTRIBUTE_labelRotation);
 		if (labelRotation != null && labelRotation.length() > 0)
-			plot.setLabelRotation(Double.valueOf(labelRotation).doubleValue());
+			plot.setLabelRotation(Double.valueOf(labelRotation));
 
 		return plot;
 	}
